Regrouping article and chart in same component for saving zoom state
