+++ src/app/mindmap/article/article.component.css
