+++ src/app/mindmap/article/article.component.html
@@ -1,6 +1,6 @@
-<div class="content" role="main">
+<div class="body-container content " role="main">
     <mat-card>
-        <button routerLink="/mind-map" mat-raised-button color="primary">
+        <button (click) = "closeArticle()" mat-raised-button color="primary">
             <mat-icon>undo</mat-icon>
             Retour 
         </button>
