+++ src/app/mindmap/article/article.component.ts
@@ -1,7 +1,9 @@
-import { Component } from '@angular/core';
+import { Component, OnInit, Input , Output, EventEmitter} from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs'
 import { ViewEncapsulation } from '@angular/core';
+import { Location } from '@angular/common';
+
 @Component({
   selector: 'app-article',
   templateUrl: './article.component.html',
@@ -9,22 +11,28 @@ import { ViewEncapsulation } from '@angular/core';
   encapsulation: ViewEncapsulation.None,
 })
 
-export class ArticleComponent {
-  private sub: Subscription;
+export class ArticleComponent  implements OnInit{
+
+  @Input() postId : string;
+  @Output() closedArticle = new EventEmitter<void>();
 
   post: string;
-  constructor(private route: ActivatedRoute) {
-  }
+  paramPostId : string; 
+
+  constructor(private route: ActivatedRoute, private location: Location) {}
 
   ngOnInit() {
-    this.sub = this.route.params.subscribe(params => {
-      this.post = './assets/articles/' +  params['articleId'] + '.md';
-    });
+    if (!this.postId) {
+      this.paramPostId = this.route.snapshot.paramMap.get('articleId');
+      this.postId = this.paramPostId;
+    }
+    this.post = './assets/articles/' +  this.postId + '.md';
   }
 
-  ngOnDestroy() {
-    if (this.sub) {
-      this.sub.unsubscribe();
-    }
+  closeArticle = () => {
+    if (this.paramPostId) {
+      this.location.back();
+    } 
+    this.closedArticle.emit()
   }
 }
