+++ src/app/mindmap/chart/chart.component.html
@@ -1,0 +1,6 @@
+<div class="componant-title">
+    <h1>Cartographie du Modern Delivery (WIP)</h1>
+</div>
+<mat-card>
+    <div #chart id="chart" class="container" (window:resize)="onResize()"></div>
+</mat-card>
