+++ src/app/mindmap/mindmap.component.html
@@ -1,8 +1,5 @@
-<div class="content" role="main">
-    <div class="componant-title">
-        <h1>Cartographie du Modern Delivery (WIP)</h1>
-    </div>
-    <mat-card>
-        <div #chart id="chart" class="container" (window:resize)="onResize()"></div>
-    </mat-card>
-</div>
+<app-chart [style.display] = "isPostOpened ? 'none' : 'block'"  
+(articleComponentOpened) = "openArticleComponent($event)" 
+class="content" role="main" (mouseover)="disableBodyScrolling()" (mouseleave)="activateBodyScrolling()"> </app-chart>
+<app-article *ngIf="isPostOpened && postIdToOpen" [postId] = "postIdToOpen"  
+(closedArticle) = "closeArticleComponent()"> </app-article>
